+++ app/src/main/kotlin/org/btcmap/auth/RequestBuilderExt.kt
@@ -2,7 +2,7 @@ package org.btcmap.auth
 
 import okhttp3.Request
 
-object PublicRequest
+internal object PublicRequest
 
 fun Request.Builder.withoutAuth(): Request.Builder = apply {
     tag(PublicRequest::class.java, PublicRequest)